chore: cleaning and renaming
