+++ src/load.py
@@ -1,103 +1,106 @@
 from __future__ import annotations
 
-import random
+import itertools
 
-import plotly.graph_objects as go
 from database.database_config import create_connection
-from flask import Flask
-from flask import render_template
-from flask import request
-from graphs.demografica_faixa_etaria_atendimento import obter_dados_distribuicao_faixa_etaria
-from graphs.demografica_faixa_etaria_problema import obter_dados_principais_problemas_faixa_etaria
-from graphs.demografica_genero_atendimento import obter_dados_distribuicao_genero
-from graphs.demografica_genero_problema import obter_dados_principais_problemas_genero
-from graphs.geografica_regiao_distribuicao import obter_dados_distribuicao_regiao
-from graphs.geografica_regiao_variacao import obter_dados_variacao_tempo
-from graphs.reclamacao_assuntos import obter_dados_assuntos_recorrentes
-from graphs.reclamacao_problema_comum import obter_dados_problemas_comuns
-from graphs.reclamacao_problema_regiao import obter_dados_destaques_regiao_uf
-from graphs.temporal_sazonalidade import obter_dados_sazonalidade
-from graphs.temporal_tendencias import obter_dados_tendencias
+from psycopg2 import Error
+from tqdm import tqdm
+from transform import transform
 
-# Conectar ao banco de dados PostgreSQL
+
+def insert_data_from_dataframe(df):
+    try:
+        conn = create_connection()
+        cursor = conn.cursor()
+
+        total_rows = len(df)
+        progress_bar = tqdm(total=total_rows)
+
+        batch_size = 50000  # Set the desired batch size
+
+        # Create batches of rows for batch insertion
+        rows_batches = [df[i:i + batch_size] for i in range(0, total_rows, batch_size)]
+
+        # Start a transaction using the context manager
+        with conn:
+            for batch in rows_batches:
+                rows = []
+                for index, row in batch.iterrows():
+                    # Create a tuple of values for each row
+                    row_data = (
+                        row['CodigoRegiao'], row['Regiao'],
+                        row['CodigoTipoAtendimento'], row['DescricaoTipoAtendimento'],
+                        row['CodigoAssunto'], row['DescricaoAssunto'], row['GrupoAssunto'],
+                        row['CodigoProblema'], row['DescricaoProblema'], row['GrupoProblema'],
+                        row['AnoAtendimento'], row['TrimestreAtendimento'], row['MesAtendimento'], row['DataAtendimento'],
+                        row['UF'], row['SexoConsumidor'], row['FaixaEtariaConsumidor'], row['CEPConsumidor']
+                    )
+                    rows.append(row_data)
+
+                # Insertion on table Regiao
+                cursor.executemany("""
+                    INSERT INTO Regiao (CodigoRegiao, Regiao)
+                    VALUES (%s, %s)
+                    ON CONFLICT (CodigoRegiao) DO NOTHING
+                """, [(row[0], row[1]) for row in rows])
+
+                # Insertion on table TipoAtendimento
+                cursor.executemany("""
+                    INSERT INTO TipoAtendimento (CodigoTipoAtendimento, DescricaoTipoAtendimento)
+                    VALUES (%s, %s)
+                    ON CONFLICT (CodigoTipoAtendimento) DO NOTHING
+                """, [(row[2], row[3]) for row in rows])
+
+                # Insertion on table Assunto
+                cursor.executemany("""
+                    INSERT INTO Assunto (CodigoAssunto, DescricaoAssunto, GrupoAssunto)
+                    VALUES (%s, %s, %s)
+                    ON CONFLICT (CodigoAssunto) DO NOTHING
+                """, [(row[4], row[5], row[6]) for row in rows])
+
+                # Insertion on table Problema
+                cursor.executemany("""
+                    INSERT INTO Problema (CodigoProblema, DescricaoProblema, GrupoProblema)
+                    VALUES (%s, %s, %s)
+                    ON CONFLICT (CodigoProblema) DO NOTHING
+                """, [(row[7], row[8], row[9]) for row in rows])
+
+                # Insertion on table Atendimento
+                cursor.executemany("""
+                    INSERT INTO Atendimento (AnoAtendimento, TrimestreAtendimento, MesAtendimento, DataAtendimento,
+                                            CodigoRegiao, UF, CodigoTipoAtendimento, CodigoAssunto, CodigoProblema,
+                                            SexoConsumidor, FaixaEtariaConsumidor, CEPConsumidor)
+                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
+                """, [(row[10], row[11], row[12], row[13], row[0], row[14], row[2], row[4], row[7], row[15], row[16], row[17]) for row in rows])
+
+                progress_bar.update(len(batch))
+
+        progress_bar.close()
+        print('Data copied successfully')
+
+        return True
+
+    except (Exception, Error) as error:
+        print('Error while executing queries:', error)
+        return False
+    finally:
+        if conn:
+            cursor.close()
+            conn.close()
+            print('Connection closed.')
+
+
+# Set the desired values for year and trimester
+year_values = range(2)
+trimester_values = range(4)
+
+# Create the connection outside the loop
 conn = create_connection()
 
-app = Flask(__name__)
+for year, trimester in itertools.product(year_values, trimester_values):
+    df = transform(year_index=year, trimester_index=trimester + 1)
+    insert_data_from_dataframe(df)
 
-
-@app.route('/')
-def index():
-    return render_template('index.html')
-
-
-@app.route('/temporal_tendencias')
-def temporal_tendencias():
-    plot = obter_dados_tendencias()
-    return render_template('temporal_tendencias.html', plot=plot)
-
-
-@app.route('/temporal_sazonalidade')
-def temporal_sazonalidade():
-    plot = obter_dados_sazonalidade()
-    return render_template('temporal_sazonalidade.html', plot=plot)
-
-
-@app.route('/geografica_regiao_distribuicao')
-def geografica_regiao_distribuicao():
-    plot = obter_dados_distribuicao_regiao()
-    return render_template('geografica_regiao_distribuicao.html', plot=plot)
-
-
-@app.route('/geografica_regiao_variacao')
-def geografica_regiao_variacao():
-    plot = obter_dados_variacao_tempo()
-    return render_template('geografica_regiao_variacao.html', plot=plot)
-
-
-@app.route('/reclamacao_assuntos')
-def reclamacao_assuntos():
-    ano_selecionado = request.args.get('ano')
-    plot = obter_dados_assuntos_recorrentes(ano_selecionado)
-    return render_template('reclamacao_assuntos.html', plot=plot)
-
-
-@app.route('/reclamacao_problema_comum')
-def reclamacao_problema_comum():
-    ano_selecionado = request.args.get('ano')
-    plot = obter_dados_problemas_comuns(ano_selecionado)
-    return render_template('reclamacao_problema_comum.html', plot=plot)
-
-
-@app.route('/reclamacao_problema_regiao')
-def reclamacao_problema_regiao():
-    uf_regiao = request.args.get('uf_regiao')
-    plot = obter_dados_destaques_regiao_uf(uf_regiao)
-    return render_template('reclamacao_problema_regiao.html', plot=plot)
-
-
-@app.route('/demografica_genero_atendimento')
-def demografica_genero_atendimento():
-    plot = obter_dados_distribuicao_genero()
-    return render_template('demografica_genero_atendimento.html', plot=plot)
-
-
-@app.route('/demografica_faixa_etaria_atendimento')
-def demografica_faixa_etaria_atendimento():
-    plot = obter_dados_distribuicao_faixa_etaria()
-    return render_template('demografica_faixa_etaria_atendimento.html', plot=plot)
-
-
-@app.route('/demografica_genero_problema')
-def demografica_genero_problema():
-    plot = obter_dados_principais_problemas_genero()
-    return render_template('demografica_genero_problema.html', plot=plot)
-
-
-@app.route('/demografica_faixa_etaria_problema')
-def demografica_faixa_etaria_problema():
-    plot = obter_dados_principais_problemas_faixa_etaria()
-    return render_template('demografica_faixa_etaria_problema.html', plot=plot)
-
-
-if __name__ == '__main__':
-    app.run(debug=True)
+# Close the connection outside the loop
+if conn:
+    conn.close()
